Fixed some duplicated code blocks and bugs.
