+++ .gitignore
@@ -4,4 +4,5 @@
 /.settings/
 *.iml
 .classpath
-.project 
+.project
+*.sh
