+++ src/main/java/me/john000708/slimexpansion/machines/XpansionContainer.java
@@ -1,7 +1,6 @@
 package me.john000708.slimexpansion.machines;
 
 import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
-import jdk.internal.jline.internal.Nullable;
 import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.InvUtils;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
@@ -13,6 +12,7 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 import me.mrCookieSlime.Slimefun.api.energy.ChargableBlock;
 import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
