+++ src/main/java/me/john000708/slimexpansion/utils/MachineUtils.java
@@ -1,0 +1,22 @@
+package me.john000708.slimexpansion.utils;
+
+import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.Item.CustomItem;
+import me.mrCookieSlime.Slimefun.api.BlockStorage;
+import org.bukkit.Bukkit;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
+import org.bukkit.inventory.Inventory;
+
+public class MachineUtils {
+    public static Inventory inject(Block b, int[] outputSlots) {
+        int size = BlockStorage.getInventory(b).toInventory().getSize();
+        Inventory inv = Bukkit.createInventory(null, size);
+        for (int i = 0; i < size; i++) {
+            inv.setItem(i, new CustomItem(Material.COMMAND_BLOCK, "&4ALL YOUR PLACEHOLDERS ARE BELONG TO US"));
+        }
+        for (int slot : outputSlots) {
+            inv.setItem(slot, BlockStorage.getInventory(b).getItemInSlot(slot));
+        }
+        return inv;
+    }
+}
